Better handle device availability
* Set common CLIENT connection availability topic
* Set LWT message to send "offline" to client availability topic
* make the device availability topic configurable
* make device availability optional
* retain "online" message

This solves a race condition where we were sending "online" before HA
was actually listening for it ("retain"). And deals with the fact that
LWT is set on connection (and can't be set later, apparently), so we
have to have some common "availability" queue that can be common for all
devices reported by the client.... except for tunnels, where we probably
just don't want to use this at all, as it doesn't make sense for
something reported from multiple clients.