Avoid unecessary stop/start of check threads
If it's already running, just let it keep running. This will prevent
unecessary flapping noise when we connect and start (assuming HA is
online if the status queue isn't set to retain)... but also reduce
noise/churn if it is and we (almost) immediately receive an "online".

We'll still re-send the discovery messages, though.